[Remove SAT] Remove sat-choice-icon.jsx (#397)
## Summary:

Previous PRs in stack:
* #394
* #395

This PR removes the `sat-choice-icon.jsx`. This was part of the Radio widget and provided a SAT-specific choice icon. Now that we're removing all SAT-specific code, we can eliminate this choice-icon.

Issue: LC-588

## Test plan:

`yarn flow`
`yarn test`

Author: jeremywiebe

Reviewers: jeremywiebe, handeyeco, Myranae

Required Reviewers:

Approved By: handeyeco

Checks: ✅ codecov/project, ✅ Lint, Flow, and Test (ubuntu-latest, 16.x), ✅ Extract i18n strings (ubuntu-latest, 16.x), ✅ Cypress Coverage (ubuntu-latest, 16.x), ✅ Check builds for changes in size (ubuntu-latest, 16.x), ✅ Check for .changeset file (ubuntu-latest, 16.x), ✅ Jest Coverage (ubuntu-latest, 16.x), ✅ gerald

Pull Request URL: https://github.com/Khan/perseus/pull/397
